Tests/LibWeb/Text/expected/form-requestSubmit.txt

Submitter is null
<INPUT id="form-associated-button" >
Exception: NotFoundError
Exception: TypeError
